requirements.txt
numpy
scikit-learn
scikit-image
scipy
pylint
autopep8
tensorflow
matplotlib
imbalanced-learn
Pillow
pyyaml
h5py
mlxtend
nltk